Add --long to version numbering.
Other lower level parts of the packaging really want to know
the number of commits and the unique sha1.  They'd be useful
too so we're merely asking the version to always give them
to us.

It would be nice to do this better, but it's a bit of a
Gordian knot at the moment.

Change-Id: Ib76806cccaf0ec28988555b54a5c38d135a28cac
Reviewed-on: http://review.couchbase.org/12200
Tested-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>