Revert "Make tests c89 friendly"
This change neglects to include macros.h in the dist file
and thus the package options fail.  I tried adding it to
Makefile.am in the sources, which seems to be what is called
for per automake documentation, but the net result was not good.

This reverts commit 9b17c40788bce8eb984e3deabaea89eb9e8b8d93.

Example of what I tried:

Change-Id: I1eafbeb3aceefe4d2ea6a862a242281fe052cdc7
Reviewed-on: http://review.couchbase.org/12372
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Matt Ingenthron <matt@couchbase.com>